Move to a lazily initialized reload-based model
Each resource now takes a snapshot of its details when it's first
created and lazily initializes its child resources. A resource's
state can be refreshed by calling the reload method.

These changes lower the number of HTTP calls that the binding makes
to the server and ensures that a resource's state is consistent
and remains constant (until reload is called).